Protecting capitalized words in titles
Capitalized words in titles. Words in titles that are to be capitalized (such as proper names) must be protected by placing the first (upper case) letter in braces.
Example: On {B}anach spaces.
Note that this applies only to title fields (title and booktitle). In other fields, such as journal names, no lower case conversion is done, so there is no need to protect capitalized words.